components: <SideNavButton /> separately + better styles
