+++ components/SideNav.js
@@ -11,6 +11,7 @@ import {
   processColor,
 } from 'react-native-web';
 import * as cfg from './config.js';
+export {SideNavButton} from './SideNavButton.js';
 
 const emptyFunction = () => {};
 
@@ -54,66 +55,6 @@ export function SideNav<Item: NavItem>({items, active, onActive, outlineColor}: 
 }
 
 SideNav.defaultProps = {
-  outlineColor: cfg.color.black,
-};
-
-type SideNavButtonProps = {
-  label: string,
-  badge?: React.Node,
-  icon?: React.Node,
-  outlineColor: string,
-  onPress: () => void,
-  active: boolean,
-};
-
-export function SideNavButton({
-  active,
-  label,
-  badge,
-  icon,
-  outlineColor,
-  onPress,
-}: SideNavButtonProps) {
-  const textStyle = {
-    color: outlineColor,
-    // fontWeight: cfg.fontWeight.semibold,
-    fontSize: active ? cfg.fontSize.base : cfg.fontSize.small,
-    fontWeight: active ? cfg.fontWeight.black : cfg.fontWeight.normal,
-  };
-  const viewStyle = {
-    height: 34,
-    flexDirection: 'row',
-    alignItems: 'center',
-    borderRadius: cfg.borderRadius.small,
-    paddingVertical: cfg.padding.size2,
-    paddingHorizontal: cfg.padding.size4,
-  };
-  const iconWrapperStyle = {
-    width: 30,
-  };
-  return (
-    <View>
-      <TouchableHighlight
-        disabled={Boolean(active)}
-        onPress={onPress}
-        underlayColor={processColor(outlineColor, 0.1)}
-        delayPressOut={0}
-        style={{
-          borderRadius: cfg.borderRadius.small,
-        }}>
-        <View style={viewStyle}>
-          <View style={iconWrapperStyle}>{icon}</View>
-          <View style={{flexGrow: 1}}>
-            <Text style={textStyle}>{label}</Text>
-          </View>
-          {badge != null && <View>{badge}</View>}
-        </View>
-      </TouchableHighlight>
-    </View>
-  );
-}
-
-SideNavButton.defaultProps = {
   outlineColor: cfg.color.black,
 };
